:rotating_light: Fix all linter warnings
