+++ common/network/include/network/network_manager.h
@@ -7,6 +7,8 @@
 */
 
 #pragma once
+
+#include <string.h>
 #include "network/network_types.h"
 
 
@@ -46,6 +48,15 @@ void network_manager_add_waiting_socket(network_manager_t *manager,
  * @param mode The mode of the socket
  */
 void network_manager_remove_waiting_socket(network_manager_t *manager,
+    int socket, waiting_socket_mode_t mode);
+/**
+ * @brief Find a waiting socket in a network manager
+ * @param manager The manager to find the socket in
+ * @param socket The socket to find
+ * @param mode The mode of the socket to find
+ * @return The index of the socket or -1 if not found
+ */
+ssize_t network_manager_find_waiting_socket(network_manager_t *manager,
     int socket, waiting_socket_mode_t mode);
 /**
  * @brief Handle all waiting sockets in a network manager
@@ -102,3 +113,18 @@ request_header_t *deserialize_request_header(char *header_str);
  * @return The byte array of the header
  */
 response_header_t *deserialize_response_header(char *header_str);
+/**
+ * @brief Iterate over all waiting sockets in a network manager
+ * and check if any of the sockets are set
+ * @param manager The manager to iterate through
+ * @param called The array of called sockets to fill
+ * @param called_count The count of called sockets
+ */
+void call_set_sockets(const network_manager_t *manager,
+    waiting_socket_t ***called, size_t *called_count);
+/**
+ * @brief Empty the read and write fds of a network manager
+ * and set them to the waiting sockets
+ * @param manager The manager to reset the fds of
+ */
+void network_manager_reset_fds(network_manager_t *manager);
