+++ common/network/src/networking.c
@@ -8,7 +8,6 @@
 
 #include "network/network_manager.h"
 #include "network/networking.h"
-#include "json/json.h"
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
@@ -26,90 +25,88 @@ api_handler_t *network_create_api_handler(host_t *host)
     return handler;
 }
 
-request_t *network_create_request(route_t route,
-    param_t params[PARAMS_MAX], json_t *body)
+static response_header_t *receive_response_header(int socket)
 {
-    request_t *request = calloc(1, sizeof(request_t));
-    char *body_str = json_serialize(body);
+    response_header_t *header = calloc(1, sizeof(response_header_t));
+    char *header_str = calloc(sizeof(response_header_t), sizeof(char));
 
-    if (!request)
+    if (header == NULL || header_str == NULL)
         return NULL;
-    request->header.route = route;
-    request->header.content_length = body_str == NULL ? 0 : strlen(body_str);
-    for (int i = 0; i < PARAMS_MAX; i++)
-        request->params[i] = params[i];
-    request->body = body_str;
-    return request;
+    recv(socket, header_str, sizeof(response_header_t), 0);
+    memcpy(header, header_str, sizeof(response_header_t));
+    free(header_str);
+    return header;
 }
 
-request_t *network_create_request_no_params(route_t route, json_t *body)
+static response_t *receive_response(response_header_t *header, int socket)
 {
-    param_t params[PARAMS_MAX] = {0};
+    response_t *response = calloc(1, sizeof(response_t));
+    char *content_str = calloc(header->content_length + sizeof(param_t) *
+        PARAMS_MAX, sizeof(char));
 
-    return network_create_request(route, params, body);
+    if (response == NULL || content_str == NULL)
+        return NULL;
+    recv(socket, content_str, header->content_length + sizeof(param_t) *
+        PARAMS_MAX, 0);
+    response->header = *header;
+    response->body = calloc(header->content_length, sizeof(char));
+    if (response->body == NULL)
+        return NULL;
+    strncpy(response->body, content_str + sizeof(param_t) * PARAMS_MAX,
+        header->content_length);
+    return response;
 }
 
-request_t *network_create_request_no_body(route_t route,
-    param_t params[PARAMS_MAX])
+static void network_receive_response_consumer(int socket, void *data)
 {
-    return network_create_request(route, params, NULL);
-}
-
-request_t *network_create_request_no_params_no_body(route_t route)
-{
-    param_t params[PARAMS_MAX] = {0};
-
-    return network_create_request(route, params, NULL);
-}
-
-static void network_receive_response_consumer(int socket, void *data) {
-    network_promise_consumer_t consumer = (network_promise_consumer_t)data;
-    response_header_t *response_header;
-    char *response_header_str = calloc(sizeof(response_header_t), sizeof(char));
-    char *response_content_str = NULL;
+    request_promises_t *promise = (request_promises_t *)data;
+    response_header_t *header = receive_response_header(socket);
     response_t *response = NULL;
 
-    if (response_header_str == NULL) {
-        close(socket);
-        return;
-    }
-    recv(socket, response_header_str, sizeof(response_header_t), 0);
-    response_header = deserialize_response_header(response_header_str);
-    response_content_str = calloc(sizeof(param_t) * PARAMS_MAX + response_header->content_length, sizeof(char));
-    if (response_content_str == NULL) {
-        close(socket);
-        free(response_header_str);
-        return;
-    }
-    recv(socket, response_content_str, sizeof(param_t) * PARAMS_MAX + response_header->content_length, 0);
-    response = deserialize_response(response_header, response_content_str);
-    consumer(response);
-    free(response_header_str);
-}
-
-static void network_send_request_consumer(int socket, void *data) {
-    request_promises_t *promise = (request_promises_t *)data;
-    char *request_str = serialize_request(promise->request);
-    waiting_socket_t *waiting_socket = malloc(sizeof(waiting_socket_t));
-
-    if (request_str == NULL || waiting_socket == NULL) {
+    if (header == NULL) {
         close(socket);
         free(promise);
-        free(waiting_socket);
         return;
     }
-    send(socket, request_str, promise->request->header.content_length + sizeof(request_header_t) + sizeof(param_t) * PARAMS_MAX, 0);
+    response = receive_response(header, socket);
+    if (response == NULL) {
+        close(socket);
+        free(header);
+        free(promise);
+        return;
+    }
+    promise->consumer(response);
+    free(header);
+    free(promise);
+    close(socket);
+}
+
+static void network_send_request_consumer(int socket, void *data)
+{
+    request_promises_t *promise = (request_promises_t *)data;
+    char *request_str = serialize_request(promise->request);
+    waiting_socket_t *w = malloc(sizeof(waiting_socket_t));
+
+    if (request_str == NULL || w == NULL) {
+        close(socket);
+        free(promise);
+        free(w);
+        return;
+    }
+    send(socket, request_str, promise->request->header.content_length +
+        sizeof(request_header_t) + sizeof(param_t) * PARAMS_MAX, 0);
     free(request_str);
-    waiting_socket->socket = socket;
-    waiting_socket->mode = READ;
-    waiting_socket->data = promise->consumer;
-    waiting_socket->consumer = network_receive_response_consumer;
-    network_manager_add_waiting_socket(promise->handler->manager, waiting_socket);
+    w->socket = socket;
+    w->mode = READ;
+    w->data = promise->consumer;
+    w->consumer = network_receive_response_consumer;
+    network_manager_add_waiting_socket(promise->handler->manager, w);
     destroy_request(promise->request);
     free(promise);
 }
 
-void network_send_request(api_handler_t *handler, request_t *request, network_promise_consumer_t consumer)
+void network_send_request(api_handler_t *handler, request_t *request,
+    network_promise_consumer_t consumer)
 {
     int socket = socket_create_client(handler->host);
     waiting_socket_t *waiting_socket;
@@ -125,12 +122,10 @@ void network_send_request(api_handler_t *handler, request_t *request, network_pr
         free(waiting_socket);
         return;
     }
-    promise->request = request;
-    promise->consumer = consumer;
-    promise->handler = handler;
-    waiting_socket->socket = socket;
-    waiting_socket->mode = WRITE;
-    waiting_socket->data = promise;
+    *(promise) = (request_promises_t){.handler = handler, .request = request,
+        .consumer = consumer};
+    *(waiting_socket) = (waiting_socket_t){.socket = socket, .mode = WRITE,
+        .data = promise};
     waiting_socket->consumer = network_send_request_consumer;
     network_manager_add_waiting_socket(handler->manager, waiting_socket);
 }
