+++ common/network/src/router_requests.c
@@ -12,45 +12,52 @@
 #include "network/router.h"
 #include "network/network_manager.h"
 
+void apply_middlewares(const network_router_t *router, request_t **request)
+{
+    route_t req_route = (*request)->header.route;
+    middlewares_t *tmp = NULL;
+
+    for (size_t j = 0; j < router->middlewares_count; j++) {
+        tmp = &router->middlewares[j];
+        if (strncmp(tmp->route->path, req_route.path,
+            strlen(tmp->route->path)) == 0
+            && tmp->route->method == req_route.method)
+            (*request) = tmp->handler((*request), tmp->data);
+    }
+}
+
+bool run_route(const network_router_t *router, request_t **request,
+    response_t **response)
+{
+    route_t req_route = (*request)->header.route;
+    route_t *tmp = NULL;
+
+    apply_middlewares(router, request);
+    for (size_t i = 0; i < router->routes_count; i++) {
+        tmp = router->routes[i].route;
+        if (strncmp(tmp->path, req_route.path, strlen(tmp->path)) == 0
+            && tmp->method == req_route.method) {
+            *response = router->routes[i].handler(*request);
+            return true;
+        }
+    }
+    return false;
+}
+
 void router_handle_request(int client_socket, void *data)
 {
-    network_router_t *router = (network_router_t *)data;
+    network_router_t *router = (network_router_t *) data;
     request_t *request = router_read_request(client_socket);
     waiting_socket_t *response_sock = malloc(sizeof(waiting_socket_t));
-    bool is_handled = false;
     response_t *response = NULL;
+    bool route_found = false;
 
     if (!request) {
-        response = calloc(1, sizeof(response_t));
-        response->header.status_code = 500;
-        response->header.content_length = strlen("{\"error\":\"Internal server error\"}");
-        response->body = strdup("{\"error\":\"Internal server error\"}");
+        response = create_server_error_response();
     } else {
-        for (size_t i = 0; i < router->routes_count; i++) {
-            if (strcmp(router->routes[i].route->path,
-                       request->header.route.path) == 0
-                && router->routes[i].route->method ==
-                   request->header.route.method) {
-                for (size_t j = 0; j < router->middlewares_count; j++) {
-                    if (strncmp(router->middlewares[j].route->path,
-                                request->header.route.path,
-                                strlen(router->middlewares[j].route->path)) == 0
-                        && router->middlewares[j].route->method ==
-                           request->header.route.method)
-                        request = router->middlewares[j].handler(request,
-                                                                 router->middlewares[j].data);
-                }
-                response = router->routes[i].handler(request);
-                is_handled = true;
-                break;
-            }
-        }
-        if (!is_handled) {
-            response = calloc(1, sizeof(response_t));
-            response->header.status_code = 404;
-            response->header.content_length = strlen("{\"error\":\"Not found\"}");
-            response->body = strdup("{\"error\":\"Not found\"}");
-        }
+        route_found = run_route(router, &request, &response);
+        if (!route_found)
+            response = create_not_found_response();
     }
     response_sock->socket = client_socket;
     response_sock->mode = WRITE;
@@ -60,24 +67,49 @@ void router_handle_request(int client_socket, void *data)
     destroy_request(request);
 }
 
-request_t *router_read_request(int client_socket)
+static request_header_t *receive_request_header(int client_socket)
 {
-    request_t *request = NULL;
-    request_header_t *header = NULL;
+    request_header_t *header = calloc(1, sizeof(request_header_t));
     char *header_buffer = calloc(1, sizeof(request_header_t));
-    char *content_buffer = NULL;
 
-    if (!header_buffer)
+    if (!header || !header_buffer)
         return NULL;
-    if (recv(client_socket, header_buffer, sizeof(request_header_t), 0) != sizeof(request_header_t))
+    if (recv(client_socket, header_buffer, sizeof(request_header_t), 0)
+        != sizeof(request_header_t))
         return NULL;
     header = deserialize_request_header(header_buffer);
-    content_buffer = calloc(header->content_length + sizeof(param_t) * PARAMS_MAX, sizeof(char));
+    free(header_buffer);
+    return header;
+}
+
+static char *receive_request_content(int client_socket,
+    request_header_t *header)
+{
+    char *content_buffer = calloc(header->content_length + sizeof(param_t)
+        * PARAMS_MAX, sizeof(char));
+
     if (!content_buffer)
         return NULL;
-    if (recv(client_socket, content_buffer, header->content_length + sizeof(param_t) * PARAMS_MAX, 0) != (ssize_t) (header->content_length + sizeof(param_t) * PARAMS_MAX))
+    if (recv(client_socket, content_buffer, header->content_length
+        + sizeof(param_t) * PARAMS_MAX, 0)
+        != (ssize_t) (header->content_length + sizeof(param_t) * PARAMS_MAX))
         return NULL;
-    request = deserialize_request(header, content_buffer);
-    free(header_buffer);
+    return content_buffer;
+}
+
+request_t *router_read_request(int client_socket)
+{
+    request_header_t *header = receive_request_header(client_socket);
+    char *content = NULL;
+    request_t *request = NULL;
+
+    if (!header)
+        return NULL;
+    content = receive_request_content(client_socket, header);
+    if (!content)
+        return NULL;
+    request = deserialize_request(header, content);
+    free(header);
+    free(content);
     return request;
 }
